RM#58656 Remove changing product name on an invoice line in qty onchange (#9677)
* RM#58656 Remove changing product name on an invoice line in qty onchange

* RM#58656 spotless

(cherry picked from commit 9967c6699a90bcda690932af76f1bc5f5de132dc)

+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/service/SupplierCatalogServiceImpl.java
@@ -237,7 +237,8 @@ public class SupplierCatalogServiceImpl implements SupplierCatalogService {
     BigDecimal minQty = this.getMinQty(product, supplierPartner, company);
 
     if (qty.compareTo(minQty) < 0) {
-      String msg = String.format(I18n.get(IExceptionMessage.PURCHASE_ORDER_LINE_MIN_QTY), minQty);
+      String msg =
+          String.format(I18n.get(IExceptionMessage.PURCHASE_ORDER_LINE_MIN_QTY), minQty.intValue());
 
       if (request.getAction().endsWith("onchange")) {
         response.setFlash(msg);
